backend/requirements.txt
sentinelhub
numpy
pyproj
matplotlib
python-dotenv
